pyproject.toml
[project]
name = "gov-health"
version = "0.1.0"
description = "Governance health analytics pipeline — Parquet extraction from yaci-store"
requires-python = ">=3.11"
dependencies = [
    "psycopg[binary]>=3.1",
    "pyarrow>=18.0",
    "click>=8.0",
    "tqdm>=4.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
duckdb = ["duckdb>=1.1"]

[project.scripts]
gov-health = "gov_health.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
